Fixing data type based on SAS specification

+++ src/readCoeData.hpp
@@ -6,4 +6,4 @@
 #include <array>
 #include <string>
 //Reading the coeficent data function will only take a file name and return an intager array to the caller of the function.
-std::vector<int16_t> readCoeData(std::string fileName);
+std::vector<float> readCoeData(std::string fileName);
